Refactor benchmark results to results/<model>/<hardware>.md
- Move existing results into dsv3_2/ subdirectory
- Update gsm8k and arithmetic bench.py to write to model-scoped paths
- Add dsv3_2_nvfp4 to MODELS in all benchmark runners
- Update CLAUDE.md to document new path scheme
- Fix LayerNorm/Hadamard dtype issues for bf16 benchmarks

+++ partitions/dsv3_2_nvfp4/ad727876/model.py
@@ -70,7 +70,7 @@ def layer_norm(
     dim: int,
     eps: float,
 ) -> torch.Tensor:
-    return F.layer_norm(x.float(), (dim,), weight, bias, eps).type_as(x)
+    return F.layer_norm(x.float(), (dim,), weight.float(), bias.float(), eps).type_as(x)
 
 
 def precompute_freqs_cis(
